Added render from pre-setup scene
+ Reorganised folder structure.
