+++ .gitignore
@@ -1,3 +1,4 @@
 Output/
+__pycache__/
 *.blend1
 *.log
