+++ Scripts/RenderFromSetup/renderSettings.json
@@ -1,0 +1,4 @@
+{
+    "camType" : "ORTHO",
+    "camRot" : "(65,0,30)"
+}
